[297]: validando com status codes HTTP

+++ app_locadora_carros/app/Http/Controllers/MarcaController.php
@@ -20,7 +20,7 @@ class MarcaController extends Controller
     public function index()
     {
         $all = $this->marca->all();
-        return $all;
+        return response()->json($all, 200);
     }
 
     /**
@@ -32,7 +32,7 @@ class MarcaController extends Controller
     public function store(Request $request)
     {
         $insert = $this->marca->create($request->all());
-        return $insert;
+        return response()->json($insert, 201);
     }
 
     /**
@@ -44,7 +44,12 @@ class MarcaController extends Controller
     public function show($id)
     {   
         $marca = $this->marca->find($id);
-        return $marca;
+        
+        if (!$marca) {
+            return response()->json(['erro' => 'Recurso pesquisado não existe.   '], 404);
+        }
+        
+        return response()->json($marca, 200);
     }
 
     /**
@@ -57,8 +62,13 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+
+        if (!$marca) {
+            return response()->json(['erro' => 'Não foi possível atualizar. Recurso não existe.'], 404);
+        }
+
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
@@ -70,7 +80,12 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+
+        if (!$marca) {
+            return response()->json(['erro' => 'Erro ao excluir. Recurso não existe.'], 404);
+        }
+
         $marca->delete();
-        return ['msg' => 'A marca foi removida com sucesso! '];
+        return response()->json(['msg' => 'A marca foi removida com sucesso! '], 200);
     }
 }
